Rewrite MMTk-specific code with WHEN_USING_MMTK
When the "#if USE_MMTK" directive and the `if (rb_mmtk_enabled_p())`
statement are used together, it will be very complicated and confusing.
We rewrite `string.c` using the convenient `WHEN_USING_MMTK`,
`WHEN_NOT_USING_MMTK` and `WHEN_USING_MMTK2` macros when possible.

This commit does not change the semantics of the existing code in
`string.c`.